Refactor the layout and figma_import crates
As a result of this change the `figma_import` crate now depends on the
`layout` crate, instead of the `layout` crate depending on
`figma_import`.

As a result, the `layout` crate now has no dependencies within the
workspace.

Code was largely organized using `taffy::prelude`'s modules.
