+++ crates/figma_import/src/toolkit_schema.rs
@@ -230,27 +230,3 @@ impl View {
         }
     }
 }
-
-/// The final result of a layout algorithm for a single taffy Node
-#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
-pub struct Layout {
-    // Relative ordering of the node. Render nodes with a higher order on top
-    // of nodes with lower order
-    pub order: u32,
-    pub width: f32,
-    pub height: f32,
-    /// The top-left corner of the node
-    pub left: f32,
-    pub top: f32,
-}
-impl Layout {
-    pub fn from_taffy_layout(l: &taffy::prelude::Layout) -> Layout {
-        Layout {
-            order: l.order,
-            width: l.size.width,
-            height: l.size.height,
-            left: l.location.x,
-            top: l.location.y,
-        }
-    }
-}
