+++ crates/layout/Cargo.toml
@@ -10,11 +10,7 @@ rust-version = "1.68"
 [lib]
 crate-type = ["rlib", "cdylib"]
 
-[features]
-fetch_layout = ["clap"]
-
 [dependencies]
-figma_import = {path="../figma_import"}
 log = "0.4"
 android_logger = "0.13.1"
 serde = { version = "1.0", features = ["derive", "rc"] }
@@ -23,9 +19,4 @@ serde_bytes = "0.11"
 bincode = "1.3"
 taffy = "0.3.18"
 lazy_static = "1.4.0"
-clap = { version = "4.5", features = ["derive"], optional = true }
 
-[[bin]]
-name = "fetch_layout"
-path = "src/bin/fetch_layout.rs"
-required-features = ["fetch_layout"]
